[WIP] Reorganizing code
