+++ src/core/Collapser.cpp
