+++ src/core/Collapser.h
