+++ src/core/Slicer.cpp
