+++ src/core/Slicer.h
