+++ src/core/Smoother.cpp
