+++ src/core/Smoother.h
