+++ src/core/SmootherTools.cpp
