+++ src/core/SmootherTools.h
