+++ src/core/Snapper.cpp
