+++ src/core/Snapper.h
