+++ src/core/SnapperOptions.h
