+++ src/core/Structurer.cpp
