+++ test/CMakeLists.txt
@@ -7,7 +7,8 @@ include_directories(${GTEST_INCLUDE_DIRS})
 
 include_directories(
 	${PROJECT_SOURCE_DIR}/src/
-	${PROJECT_SOURCE_DIR}/src/tessellator/
+	${PROJECT_SOURCE_DIR}/src/core/
+	${PROJECT_SOURCE_DIR}/src/drivers/
 	${PROJECT_SOURCE_DIR}/src/utils/
 	${PROJECT_SOURCE_DIR}/src/types/
 	${PROJECT_SOURCE_DIR}/test/
@@ -27,20 +28,18 @@ if (TESSELLATOR_USE_CGAL)
 		"cgal/ManifolderTest.cpp"
 		"cgal/PolyhedronToolsTest.cpp"
 		"cgal/RepairerTest.cpp"
-		"filler/FillerTest.cpp"
-		"filler/FillerToolsTest.cpp"
+		"cgal/filler/FillerTest.cpp"
+		"cgal/filler/FillerToolsTest.cpp"
 	)
 endif()
 
 add_executable(tessellator_tests 
-	"tessellator/CollapserTest.cpp"
-	"tessellator/DriverTest.cpp"
-	"tessellator/SlicerTest.cpp"
-	"tessellator/SmootherTest.cpp"
-	"tessellator/SmootherToolsTest.cpp"
-	"tessellator/SnapperTest.cpp"
-    "tessellator/StructuredDriverTest.cpp"
-    "tessellator/StructurerTest.cpp"
+	"core/CollapserTest.cpp"
+	"core/SlicerTest.cpp"
+	"core/SmootherTest.cpp"
+	"core/SmootherToolsTest.cpp"
+	"core/SnapperTest.cpp"
+    "core/StructurerTest.cpp"
 	"types/MeshTest.cpp"
 	"utils/CleanerTest.cpp"
 	"utils/CoordGraphTest.cpp"
@@ -48,16 +47,18 @@ add_executable(tessellator_tests
 	"utils/GeometryTest.cpp"
 	"utils/GridToolsTest.cpp"
 	"utils/MeshToolsTest.cpp"
+    "drivers/StructuredDriverTest.cpp"
 )
 
 target_link_libraries(tessellator_tests	
-	tessellator
+	tessellator-drivers
     GTest::gtest 
 	GTest::gtest_main
 )
 
 if (TESSELLATOR_USE_CGAL)
-	target_link_libraries(tessellator_tests tessellator_cgal_tests GTest::gtest)
+	add_executable(tessellator_tests "drivers/DriverTest.cpp")
+	target_link_libraries(tessellator_tests tessellator_cgal_tests)
 endif()
 
 add_test(tessellator ${CMAKE_RUNTIME_OUTPUT_DIRECTORY}/tessellator_tests)
