+++ test/core/SlicerTest.cpp
